Converted generic_map_004 to use move_token base rule.
