+++ docs/generic_map_rules.rst
@@ -84,7 +84,11 @@ generic_map_004
 
 |phase_1| |error| |structure|
 
-This rule checks for the closing parenthesis *)* on generic maps are on their own line.
+This rule checks the location of the closing ")" character for the generic map.
+
+The default location is on a line by itself.
+
+Refer to the section `Configuring Closing Parenthesis Location<configuring-closing-parenthesis-location>`_ for information on options.
 
 **Violation**
