+++ vsg/rules/generic_map/rule_004.py
@@ -1,18 +1,18 @@
 
-from vsg.rules import split_line_at_token_when_between_tokens
+from vsg.rules import move_token as Rule
 
 from vsg import token
 
-lTokens = []
-lTokens.append(token.generic_map_aspect.close_parenthesis)
-
-oStart = token.component_instantiation_statement.label_colon
-oEnd = token.component_instantiation_statement.semicolon
+oToken = token.generic_map_aspect.close_parenthesis
 
 
-class rule_004(split_line_at_token_when_between_tokens):
+class rule_004(Rule):
     '''
-    This rule checks for the closing parenthesis *)* on generic maps are on their own line.
+    This rule checks the location of the closing ")" character for the generic map.
+
+    The default location is on a line by itself.
+
+    Refer to the section `Configuring Closing Parenthesis Location<configuring-closing-parenthesis-location>`_ for information on options.
 
     **Violation**
 
@@ -35,5 +35,4 @@ class rule_004(split_line_at_token_when_between_tokens):
     '''
 
     def __init__(self):
-        split_line_at_token_when_between_tokens.__init__(self, 'generic_map', '004', lTokens, oStart, oEnd)
-        self.solution = 'Place closing ) on it\'s own line.'
+        Rule.__init__(self, 'generic_map', '004', oToken)
